7.53.0 Import (#737)
* added real icons

* fix color of loaded svg files

* recursively parse SVG for color

* using EventListenerList where possible

* adding Import button and some icons

* Everywhere use Color instead of ColorRGB

* more icons

* updated test results to include alpha values

* 7.53.0

* added PulseCMYK

* 7.53.1 missing string

+++ src/main/java/com/marginallyclever/makelangelo/makeart/TransformedImage.java
@@ -99,8 +99,8 @@ public class TransformedImage {
 	 * @return greyscale intensity in this region. [0...255]v
 	 */
 	public int sample(double x0, double y0, double x1, double y1) {
-		float sampleValue = 0;
-		float weightedSum = 0;
+		double sampleValue = 0;
+		double weightedSum = 0;
 
 		int left   = (int)Math.floor(x0);
 		int right  = (int)Math.ceil (x1);
